Add a partial index on account move line
An log analysis showed that the normalized query below was executed very often
with a slow explain plan using a seq scan.

  SELECT move_id, date
  FROM account_move_line
  WHERE journal_id = 0
  AND period_id = 0
  AND create_uid = 0
  AND state = ''
  ORDER BY id DESC LIMIT 0;

This query is called in the _default_get of account_move_line to find the last
unbalanced move line. The state is always 'draft' and the fields always
journal_id, period_id and create_uid so it is a good candidate for a partial
index with combined fields.

The explain plan shows an impressive improvement, dropping from ~800ms to
~0.030 ms for a table with 3906500 records. This query was called 7700 times in
a working day for a cumulated duration of 25m32s.